Added five files
Thus through this project our aim is to use the concepts of Data Visualisation and to apply them in order to better analyse these events and deduce important conclusions that will help towards better prevention of these fires by better understanding the more influential factors which lead to these events and the roles played by them. We also aim to generate sufficient awareness in the whole world through this paper about the intensity of these events and their catastrophic effects on the world and gather the attention required towards it.